Update build and add several tools and improvements (#611)
- Updated appveyor with the new log library
and also added a zip files with all artifacts.

- Updated travis to build release and debug
versons, and also changed the ubuntu version
to bionic to use new gcc, also added the new
log library.

- Improved cmake file to enable more warnings,
to have a release and debug version, and to
generate the documentation when enabled.

- Updated docker files to install the
new log library.

- Added new account and players to docker
test server. Changed god account password
to 'god'

- Update all files headers to insert
the Copyright.

- Update all .h header files to use the
Google Style.

- Fixed all 'shadow' warning about variables
with the same name.

+++ src/networkmessage.cpp
@@ -1,4 +1,6 @@
 /**
+ * @file networkmessage.cpp
+ * 
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2019 Mark Samman <mark.samman@gmail.com>
  *
@@ -26,7 +28,7 @@
 
 int32_t NetworkMessage::decodeHeader()
 {
-	int32_t newSize = static_cast<int32_t>(buffer[0] | buffer[1] << 8);
+	int32_t newSize = buffer[0] | buffer[1] << 8;
 	info.length = newSize;
 	return info.length;
 }
